# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Data Analyst Agent API
  version: 1.0.0
  description: |
    AI Data Analyst Agent 前后端共享契约。
    具体通用规则见 API_INTEGRATION_STANDARD.md。
  contact:
    name: AI Data Analyst Agent 项目组
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发
  - url: /api/v1
    description: 同源部署
...
security:
  - bearerAuth: []
paths:
...
  /system/capabilities:
    get:
      tags: [System]
      operationId: getCapabilities
      summary: 获取部署能力与限制
      responses:
        "200":
          description: 当前部署能力
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemCapabilities"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/dataset-versions/{version_id}/schema:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/VersionId"
    get:
      tags: [Schema]
      operationId: getDatasetSchema
      summary: 获取字段 Schema
      responses:
        "200":
          description: 字段定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetSchema"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Schema]
      operationId: updateDatasetSchema
      summary: 提交字段覆盖
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemaPatch"
      responses:
        "200":
          description: 更新后的字段定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetSchema"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: datatrace_session
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^prj_[A-Za-z0-9]+$"
...
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^dsv_[A-Za-z0-9]+$"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源 revision，格式为双引号包围的整数
      schema:
        type: string
        example: '"3"'
  responses:
...
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在或无权访问
              request_id: req_01JABC
              retryable: false
              details: {}
...
    PreconditionFailed:
      description: 客户端资源版本已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VERSION_CONFLICT
              message: 资源已被修改，请刷新后重试
              request_id: req_01JABC
              retryable: false
              details:
                current_revision: 4
    ValidationError:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SystemCapabilities:
      type: object
      required:
        - api_version
        - supported_file_types
        - max_upload_bytes
        - natural_language_analysis
        - auth_enabled
      properties:
        api_version:
          type: string
          example: v1
        supported_file_types:
          type: array
          items:
            type: string
            enum: [csv, xls, xlsx, parquet]
        max_upload_bytes:
          type: integer
          minimum: 1
          example: 524288000
        natural_language_analysis:
          type: boolean
        auth_enabled:
          type: boolean
        polling:
          $ref: "#/components/schemas/PollingPolicy"
...
    PollingPolicy:
      type: object
      required: [initial_interval_ms, steady_interval_ms, background_interval_ms]
      properties:
        initial_interval_ms:
          type: integer
          example: 1000
        steady_interval_ms:
          type: integer
          example: 3000
        background_interval_ms:
          type: integer
          example: 10000
...
    DatasetSchema:
      type: object
      required: [dataset_version_id, revision, columns, low_confidence_count]
      properties:
        dataset_version_id:
          type: string
        revision:
          type: integer
          minimum: 1
        low_confidence_count:
          type: integer
          minimum: 0
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnSchema"
    ColumnSchema:
      type: object
      required:
        - name
        - physical_type
        - semantic_type
        - analysis_role
        - confidence
        - evidence
        - sensitive
      properties:
        name:
          type: string
        physical_type:
          type: string
          enum: [integer, float, string, boolean, datetime, binary, unknown]
        semantic_type:
          type: string
          enum:
            - numeric
            - categorical
            - datetime
            - boolean
            - identifier
            - text
            - currency
            - percentage
            - ordinal
            - geographic_code
            - unknown
        analysis_role:
          type: string
          enum: [feature, target, entity_key, time, identifier, text, excluded, undecided]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        evidence:
          type: array
          items:
            type: string
        date_format:
          type: [string, "null"]
        ordinal_values:
          type: [array, "null"]
          items:
            type: string
        sensitive:
          type: boolean
        user_confirmed:
          type: boolean
          default: false
    SchemaPatch:
      type: object
      additionalProperties: false
      required: [changes]
      properties:
        changes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ColumnSchemaPatch"
        reason:
          type: [string, "null"]
          maxLength: 1000
    ColumnSchemaPatch:
      type: object
      additionalProperties: false
      required: [column]
      properties:
        column:
          type: string
        semantic_type:
          type: string
          enum:
            - numeric
            - categorical
            - datetime
            - boolean
            - identifier
            - text
            - currency
            - percentage
            - ordinal
            - geographic_code
            - unknown
        analysis_role:
          type: string
          enum: [feature, target, entity_key, time, identifier, text, excluded, undecided]
        date_format:
          type: [string, "null"]
        ordinal_values:
          type: [array, "null"]
          items:
            type: string
        sensitive:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable, details]
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true